smarten then timeout in sensor data producer
the timeout was based on number of ticks instead of actual time.
So when the collection was fast (lots of ticks a second), then
the timeout would be too short, and the data producer would give
up on the sensor device before it had a chance to send any data.

git-svn-id: https://svn.concord.org/svn/projects/trunk/common/java/sensor/sensor@25474 6e01202a-0783-4428-890a-84243c50cc2b